airnest_backend/backend/requirements.txt
django==5.1.4
psycopg2-binary==2.9.6
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
channels==4.0.0
channels-redis>=4.2
daphne==4.0.0
pytz==2024.1
requests==2.31.0
django-cleanup==8.0.0
google-cloud-translate==3.15.3
dj-database-url>=2.2.0
boto3>=1.34.0
django-storages>=1.14.3